feat(projectStructure): move error response into its own package
