+++ sante.api/src/main/java/org/aksw/sante/api/exception/handler/JsonParseExceptionHandler.java
@@ -2,7 +2,7 @@ package org.aksw.sante.api.exception.handler;
 
 import com.fasterxml.jackson.core.JsonParseException;
 import org.aksw.sante.api.exception.error.ApiError;
-import org.aksw.sante.api.exception.error.ApiErrorResponse;
+import org.aksw.sante.api.exception.response.ApiErrorResponse;
 import org.aksw.sante.api.exception.error.JsonParseError;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
