+++ sante.api/src/main/java/org/aksw/sante/api/exception/handler/SearchSuggestExceptionHandler.java
@@ -2,7 +2,7 @@ package org.aksw.sante.api.exception.handler;
 
 import org.aksw.sante.api.exception.SearchSuggestException;
 import org.aksw.sante.api.exception.error.ApiError;
-import org.aksw.sante.api.exception.error.ApiErrorResponse;
+import org.aksw.sante.api.exception.response.ApiErrorResponse;
 import org.aksw.sante.api.exception.error.SearchSuggestError;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
